xjmall/requirements.txt
Django==1.11.16
requests==2.20.1
django_filter==1.1.0
django_filters==0.2.1
djangorestframework==3.9.1
django_crontab